Change player pasts and highscores to use frames instead of times (in preparation for replays).

+++ core/src/main/java/com/destroflyer/escapeloop/states/models/HighscoreDto.java
@@ -9,5 +9,5 @@ import lombok.NoArgsConstructor;
 public class HighscoreDto {
     private String context;
     private UserDto user;
-    private long score;
+    private int score;
 }
